host networking, ssh port, server paramter
The start.sh script starts the container with host networking
as default. With bridged networking does nfs export from
the container does not work, due to conflicts in the rpc
port 111

The ssh port can be configured ad container start with
SSH_PORT container environment variable. This allows to
login to the container it is started with host networking.
ssh -p $SSH_PORT localhost

Bug-fix : The start-up script forwards the https port
to the server now.

handling SIGTERM added to docker entrypoint to stop nfs-server

A signal handler for SIGTERM has been added to the docker
entrypoint script.
The handler stops the nfs-server to avoid the hanging nfs
processed after stopping the container.
The handler exits with 0, which speeds up the container
stopping.

Change-Id: I8c5f31f2d0bc1b35876b0d42f0d9390dc3375e62
Signed-off-by: Ralf Mueller <ralf.1.mueller@nokia.com>