FIX: Allow configuration of IPMI privilege level
When invoking the ipmitool command, if an IPMI privilege level is not
specified, the default is 'ADMINISTRATOR'.  In installations where the
IPMI user does not have this privilege level, this will cause IPMI
requests to fail.  This change updates the remote installer's
invocations of ipmitool to pass a configured privilege level.  If a
privilege level is not supplied in the host configuration, the code
defaults to 'ADMINISTRATOR' to mirror the ipmitool behavior.

Previous changes added this functionality to the config-manager
components; this change brings the remote installer into line
with those components.

signed-off-by: dave kormann <davek@research.att.com>
Change-Id: I752e53e0289b919430c2478bb1303f85e70f80ae