Update patch set 1
Patch Set 1:

It seems the remote installer invokes ipmitool itself, in addition to indirectly invoking it via the configuration manager.  So, as with the previous change that passed an ipmi priv_level to the configuration management APIs, we need to pass it on the remote installer's use of ipmitool as well.  This change updates the bmctools base class and its hardware subclasses to support that.

Patch-set: 1