Update patch set 1
Patch Set 1:

maybe my Python is lacking, but isn't it the case that the default value is only used if the input parameter was not defined by the caller?
if yes, in your case as you said it is always defined, so default ADMINISTRATOR does not override
but generally speaking IMO it is still good practice to leave the def value, if the internals of the function depend on it. future callers might forget to define it

Patch-set: 1