Update patch set 1
Patch Set 1:

really, the evidence is that it's MY python that's lacking, given the quantity of silly errors i've introduced...

the problem here is that python syntax requires that if a method argument has a default value, all subsequent arguments _also_ have defaults — this is to resolve ambiguity in optional-versus-required args.

there are two options we could take that would continue to supply a default: one would be to move the priv_level to the end of the argument list, the other would be to provide defaults for login_user and login_password.  I personally don't like moving the priv_level to the end of the list — it seems like that splits up the IPMI arguments confusingly.  on the other hand, I'm not sure what meaningful defaults would be for login_user and login_pass.

i don't think there's a real danger of future callers _forgetting_ to supply the value as, at present, it's required: it'd be a syntax error to supply too few arguments to the method.  we could have the initializer check if a null (or non-valid) privilege level is passed in and set it to 'ADMINISTRATOR' in that case, of course, if that'd be better.

Patch-set: 1
Reviewer: Gerrit User 180 <180@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Verified=0