Update patch set 2
Patch Set 2:

Could it be that the rpm build part is actually obsolete?
It seems that this project started with ta/ rpm boilerplate, but in the end it is used outside of TA, on the Regional Controller and pulled from a docker registry.

Patch-set: 2
Reviewer: Gerrit User 251 <251@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Verified=0