Fix: First boot from floppy, not CD-ROM
The remote installer sets the next boot to the floppy device, apparently
to force a boot from remote media.  On current OE19 firmware, this
appears to result in a boot from the default hard drive instead.  This
change sets the next boot to cd, which appears to work when issued
manually

signed-off-by: dave kormann    <davek@research.att.com>
Change-Id: Ib44550055596cf69a688b0e57060d8315cb6e887